Support mdata variable for /root/.ssh/authorized_keys file
This script creates a .ssh folder with minimal permissions, if the directory already exists the permissions will be changed to 700. If the mdata variable `root_authorized_keys` the `authorized_keys` is cleared and a header is added. I also fix the file permissions to be sure we're using the minimal permissions required for `authorized_keys`.